moved filtering logic into filter component
